+++ ChangeLog.md
@@ -4,6 +4,8 @@
 * Improve output for package index updates [#227](https://github.com/commercialhaskell/stack/issues/227)
 * Automatically update indices as necessary [#227](https://github.com/commercialhaskell/stack/issues/227)
 * --verbose flag [#217](https://github.com/commercialhaskell/stack/issues/217)
+* Remove packages (HTTPS and Git) [#199](https://github.com/commercialhaskell/stack/issues/199)
+* Config values for system-ghc and install-ghc
 
 ## 0.0.1